Fix outer run loop randomisation bug
The outer loop can be handed a list of randomisation seeds, each of
which should be used for the corresponding outer loop index. This seed
was being mistakenly set after subjects were selected from the total
subject pool, meaning every outer loop would have the same selection of
subjects when the script was run for each outer loop iteration
separately (i.e. for execution on the HPC).